test: move the change event unit tests to a separate file (#5632)
